Refactor for Redux; and add URL visits, and bookmark indication
Prior to this commit, Flux was not considered as state management
seemed fairly trivial. This commit attempted to "resolve" the
`referringVisitId` from a visit to an actual `HistoryItem`, but
that turned out to be more complex since it requires fetching all
`VisitItem` first. We do not get all `VisitItem` first since they
are a child of URL (retrievable via `HistoryItem`). There is no API
to fetch a history or visit from a `referringVisitId`.